feat: implement viral loyalty widget growth loop tests
Implemented Playwright E2E tests for the Viral Loyalty Widget growth loop in `src/ui/next/src/e2e/viral_loyalty_widget.spec.ts`.
The tests:
- Execute a real authentication flow.
- Navigate to the widget from the dashboard.
- Generate a loyalty program and verify the UI state changes (stamps fill up, button disables, etc.).
- Verify the generated share link and the clipboard copy functionality.
- Verify the mobile-first responsive layout (at 375px width).

Mocks are not used. All tests run against the authentic application flow.
Also updated `BUILD.bazel` to include this test in the Playwright suite.
