+++ src/ui/next/BUILD.bazel
@@ -7,6 +7,7 @@ exports_files(glob([
     "package.json",
     "e2e/*.spec.ts",
     "src/e2e/**/*.spec.ts",
+    "src/e2e/**/*.spec.ts",
     "src/e2e/fixtures/*",
 ]))
 
@@ -14,6 +15,7 @@ filegroup(
     name = "next_e2e_specs",
     srcs = glob([
         "e2e/*.spec.ts",
+    "src/e2e/**/*.spec.ts",
         "src/e2e/**/*.spec.ts",
         "src/e2e/fixtures/*",
     ]),
