+++ src/ui/next/src/e2e/viral_loyalty_widget.spec.ts
@@ -1,0 +1,93 @@
+import { test, expect } from './fixtures';
+
+test.describe('Viral Loyalty Widget Growth Loop', () => {
+  test('should generate a loyalty program within an authenticated session', async ({
+    page,
+    workerContext,
+  }) => {
+    // Navigate to the login page first
+    await page.goto('/login');
+
+    // Perform login (assuming standard test credentials)
+    await page.fill('input[name="email"]', 'starter@example.com');
+    await page.fill('input[name="password"]', 'password123');
+    await page.click('button[type="submit"]');
+
+    // Wait for navigation to complete
+    await page.waitForURL('**/dashboard.html');
+    await expect(page.locator('text=OHC E2E Bakery POS')).toBeVisible({ timeout: 10000 });
+
+    // Navigate to the widget from the authenticated state
+    await page.goto('/ui/viral-loyalty-widget.html');
+
+    // Validate page loaded
+    await expect(page.locator('h1')).toHaveText('Viral Loyalty Widget Generator');
+
+    // Check initial stamps
+    const emptyStamps = page.locator('.stamp.empty');
+    await expect(emptyStamps).toHaveCount(4);
+
+    const generateBtn = page.locator('#generate-btn');
+    await expect(generateBtn).toBeVisible();
+
+    // Click to generate
+    await generateBtn.click();
+    await expect(generateBtn).toBeDisabled();
+    await expect(generateBtn).toHaveText('Generating...');
+
+    // Wait for the actual API call to complete and render the result
+    const resultArea = page.locator('#result-area');
+    await expect(resultArea).toBeVisible({ timeout: 10000 });
+
+    const filledStamps = page.locator('.stamp.filled');
+    await expect(filledStamps).toHaveCount(4);
+
+    const shareLink = page.locator('#share-link');
+    // Ensure the link is actually generated by verifying it is not empty
+    await expect(shareLink).not.toBeEmpty();
+
+    // Verify copy button works without mocking clipboard
+    const copyBtn = page.locator('#copy-btn');
+    await expect(copyBtn).toHaveText('Copy');
+
+    // Give permissions for clipboard operations
+    await page.context().grantPermissions(['clipboard-read', 'clipboard-write']);
+    await copyBtn.click();
+    await expect(copyBtn).toHaveText('Copied!', { timeout: 3000 });
+
+    try {
+        const clipboardText = await page.evaluate(async () => {
+            return await navigator.clipboard.readText();
+        });
+        expect(clipboardText.length).toBeGreaterThan(0);
+    } catch (e) {
+        console.warn('Clipboard read failed (expected in some headless environments): ', e);
+    }
+
+    // Verify the back link returns to dashboard
+    const backLink = page.locator('.back-link');
+    await expect(backLink).toBeVisible();
+    await expect(backLink).toHaveAttribute('href', '/dashboard.html');
+    await backLink.click();
+    await page.waitForURL('**/dashboard.html');
+  });
+
+  test('should show responsive layout on mobile viewport', async ({ page }) => {
+    // Set viewport to a mobile layout (375px)
+    await page.setViewportSize({ width: 375, height: 812 });
+
+    await page.goto('/login');
+    await page.fill('input[name="email"]', 'starter@example.com');
+    await page.fill('input[name="password"]', 'password123');
+    await page.click('button[type="submit"]');
+    await page.waitForURL('**/dashboard.html');
+
+    await page.goto('/ui/viral-loyalty-widget.html');
+    await page.waitForTimeout(100);
+
+    // Verify Mobile layout
+    const container = page.locator('.container');
+    const box = await container.boundingBox();
+    expect(box?.width).toBeLessThanOrEqual(375);
+  });
+});
